+++ parachain/pallets/ethereum-beacon-client/fuzz/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod impls;
 pub mod types;
